pyproject.toml
[tool.poetry]
name = "slac-auth"
version = "0.1.0"
description = ""
authors = ["tanlda <ledinhanhtan.stack@gmail.com>"]
readme = "README.md"
packages = [{ include = "scripts", from = "." }]

[tool.poetry.scripts]
makemigrations = "scripts.django:makemigrations"
migrate = "scripts.django:migrate"
shell = "scripts.django:shell"
server = "scripts.django:runserver"
gunicorn = "scripts.django:gunicorn"


[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
asttokens = "2.4.1"
build = "1.0.3"
cachecontrol = "0.13.1"
certifi = "2024.2.2"
cffi = "1.16.0"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
crashtest = "0.4.1"
cryptography = "42.0.3"
decorator = "5.1.1"
defusedxml = "0.8.0rc2"
distlib = "0.3.8"
django = "5.0.2"
django-authtools = "2.0.0"
django-cors-headers = "4.3.1"
django-environ = "0.11.2"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
dulwich = "0.21.7"
executing = "2.0.1"
factory-boy = "3.3.0"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
gunicorn = "21.2.0"
identify = "2.5.35"
idna = "3.6"
importlib-metadata = "7.0.1"
iniconfig = "2.0.0"
installer = "0.7.0"
ipython = "8.21.0"
jaraco-classes = "3.3.1"
jedi = "0.19.1"
jeepney = "0.8.0"
keyring = "24.3.0"
matplotlib-inline = "0.1.6"
more-itertools = "10.2.0"
msgpack = "1.0.7"
nodeenv = "1.8.0"
oauthlib = "3.2.2"
packaging = "23.2"
parso = "0.8.3"
pexpect = "4.9.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
pluggy = "1.4.0"
pre-commit = "3.6.2"
prompt-toolkit = "3.0.43"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pycparser = "2.21"
pygments = "2.17.2"
pyjwt = "2.8.0"
pyproject-hooks = "1.0.0"
pytest-django = "4.8.0"
pytest-env = "1.1.3"
pytest-mock = "3.12.0"
python-dateutil = "2.8.2"
python3-openid = "3.2.0"
pytz = "2024.1"
pyyaml = "6.0.1"
rapidfuzz = "3.6.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
requests-toolbelt = "1.0.0"
secretstorage = "3.3.3"
shellingham = "1.5.4"
six = "1.16.0"
social-auth-app-django = "5.4.0"
social-auth-core = "4.5.3"
split-settings = "1.0.0"
sqlparse = "0.4.4"
stack-data = "0.6.3"
tomlkit = "0.12.3"
traitlets = "5.14.1"
trove-classifiers = "2024.1.31"
urllib3 = "2.2.1"
validators = "0.22.0"
virtualenv = "20.25.0"
wcwidth = "0.2.13"
zipp = "3.17.0"
shortuuid = "^1.0.13"
pytest = "^8.1.1"
faker = "^24.3.0"
microslac = "^0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120


[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__"]
reportMissingImports = true
reportMissingTypeStubs = false
reportGeneralTypeIssues = false
reportOptionalSubscript = false
reportPrivateImportUsage = "information"
reportUnusedFunction = true

pythonPlatform = "Linux"
pythonVersion = "3.11"

[tool.isort]
known_local_folder = ["webmaster_verification", "sk*"]
known_typing_extensions = "typing_extensions"
line_length = 120
no_lines_before = "TYPING_EXTENSIONS"
profile = "black"
sections = [
    "FUTURE",
    "STDLIB",
    "TYPING_EXTENSIONS",
    "THIRDPARTY",
    "LOCALFOLDER",
    "FIRSTPARTY",
]
